Haven't commited in a while now.
Biggest change is with the Chess engine. So far i have it initializing the board
with the pieces at their starting positions, printing to console, correct orientation
for each player, player input with arbitrary piece movement, and endgame when a king dies.

Whats left to do (that I plan on doing initially) is to establish the rules for piece
movement, writing the testMove functions for each of the piece types. after that is should
be a functioning chess implementation.

Possible todo:  create a score counter based on established piece value.