Progress with Chess:
Started and finished the first of the six piece move validator functions, Pawns.
Discovered and fixed (with the help of Spencer Hawkins) a terrible terrible
terrible game breaking bug with scanf (hate scanf, it can go die) in which the second
time it was called it wouldn't read the input correctly. it was solved by placeing a
space in the begining of the scanf argument string such that scanf("%c%c %c%c") became
scanf(" %c%c %c%c"). needless to say, i raged.

I say this again for good measure: scanf can go suck a dick.

There is a bright side to this stupidness however. i forsee this to be the last major bug
relating to user input (crosses fingers), which leave only the five remaining piece
check functions as the remaining hurdles.

Yippee.